inscription + connexion utilisateur dans la bdd

+++ src/main/java/com/narbaniki/timescrimble/PartieRepository.java
@@ -2,6 +2,6 @@ package com.narbaniki.timescrimble;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PartieRepository extends JpaRepository<Partie, Long>{
+public interface PartieRepository extends JpaRepository<Partie, int>{
     
 }
